Moved add_token command
