+++ ext/ci.py
@@ -1,6 +1,8 @@
 # Import our libraries
 import copy
+import discord
 from cog import Cog
+from discord.ext import commands
 from exceptions import NoTokenSet
 
 
@@ -8,12 +10,16 @@ class ContinuousIntegration(Cog):
     """
     A cog for accessing common Cotinuous integration APIs.
     """
-    def __init__(self, bot, name):
+    def __init__(self, bot, name: str, auth: str, headers: dict, endpoints: dict):
         # Save our bot for later use
         self.bot = bot
         # Save the collections
         self.tokens = bot.database[f"{name}_tokens"]
         self.picks = bot.database[f"{name}_picks"]
+        # Save our default set of parameters
+        self.auth = auth
+        self.headers = headers
+        self.endpoints = endpoints
 
     async def dump_data(self, ctx):
         """
@@ -42,14 +48,14 @@ class ContinuousIntegration(Cog):
         self.picks.delete_many({"_id": ctx.author.id})
         return True
 
-    async def generate_headers(self, ctx, default: dict, auth_name: str):
+    async def generate_headers(self, ctx, auth_name: str):
         """
         Generates a set of headers for the use on aiohttp requests.
         """
         # Get the current user token
         token = (await self.get_token(ctx.author.id))["token"]
         # Create a copy of the default haders
-        headers = copy.deepcopy(default)
+        headers = copy.deepcopy(self.headers)
         # Set the token specified by the user
         headers["Authorization"] = f"{auth_name} {token}"
         # Finally, return the headers
@@ -66,6 +72,40 @@ class ContinuousIntegration(Cog):
             raise NoTokenSet("A Travis CI.com token is required.")
         # Otherwise, return found key
         return existing
+
+    @commands.command()
+    async def addtoken(self, ctx, token: str):
+        """
+        Adds a Personal Token to our database.
+        Please note that we check the validity of the tokens prior to storing them.
+        """
+        # Send a typing
+        await ctx.trigger_typing()
+
+        # Create a copy of the default haders
+        headers = copy.deepcopy(self.headers)
+        # Set the token specified by the user
+        headers["Authorization"] = f"{self.auth} {token}"
+
+        # Request the user data
+        async with self.bot.session.get(self.endpoints["validity"], headers=headers) as resp:
+            # If the code is 403
+            if resp.status == 403:
+                await ctx.send("The token that has been specified is not valid.")
+            # If the code is 200
+            elif resp.status == 200:
+                # Update an item and create it if is not present
+                await self.tokens.replace_one({"_id": ctx.author.id}, {"_id": ctx.author.id, "token": token}, True)
+                # Notify the user
+                await ctx.send("Your token has been updated!")
+            # If the code is anything else
+            else:
+                await ctx.send(f"Error while checking for your token: Code {resp.status}")
+
+        # If the user posted on a public text channel
+        if isinstance(ctx.channel, discord.TextChannel):
+            # Delete the original message
+            await ctx.message.delete()
 
 
 def setup(bot):
